[back] 댓글, 대댓글 api 구현 - 댓글 CRUD api 구현 - 대댓글 CRUD api 구현 - 대댓글 조회의 경우 댓글에서 articleId로 일괄 처리하는 로직으로 구현

+++ BackEnd/chartflow/src/main/java/com/ssafy/chartflow/board/dto/request/RequestWriteCommentDto.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class RequestWriteCommentDto {
-    private int userId;
-    private int articleId;
+    private long userId;
+    private long articleId;
     private String content;
 }
